Fixed casting issue.

+++ src/main/java/micdoodle8/mods/galacticraft/core/items/ItemOxygenTank.java
@@ -19,9 +19,8 @@ public class ItemOxygenTank extends Item
     {
         super();
         this.setMaxStackSize(1);
-	double factor = 2.0;
-	// Config modifier goes here if anyone wants it.
-        this.setMaxDamage(Math.round(Math.pow(factor, tier - 1) * 10) * 100);
+        double factor = 2.0;
+        this.setMaxDamage((int) (Math.round(Math.pow(factor, tier - 1) * 10) * 100));
         this.setUnlocalizedName(assetName);
         this.setTextureName(GalacticraftCore.TEXTURE_PREFIX + assetName);
         this.setNoRepair();
